pkg/tasks/cleanup_backup_service_test.go
package tasks

import (
	"context"
	"github.com/go-pg/pg/v10"
	"github.com/ottogroup/penelope/pkg/http/mock"
	"github.com/ottogroup/penelope/pkg/repository"
	"github.com/ottogroup/penelope/pkg/secret"
	service2 "github.com/ottogroup/penelope/pkg/service"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
	"strings"
	"testing"
	"time"
)

const (
	cleanupServiceJobID    = "cleanup-uuid-1234"
	cleanupServiceBackupID = "cleanup-uuid-5678"
)

func TestCleanupExpiredSinkService_WithoutValidJob(t *testing.T) {
	ctx := context.Background()
	service, err := newCleanupExpiredSinkService(ctx, nil, secret.NewEnvSecretProvider())
	require.NoError(t, err)
	service.scheduleProcessor = MockScheduleProcessor{
		shouldReturnValidJob:    false,
		shouldReturnValidBackup: false,
		ctx:                     ctx,
	}

	_, stdErr, err := captureStderr(func() {
		service.Run(ctx)
	})

	require.NoError(t, err)
	logMsg := "could not get list of expired backup backups for backup type BigQuery"
	assert.Containsf(t, strings.TrimSpace(stdErr), logMsg, "Run should write log message %q but it logged\n\t%s", logMsg, stdErr)
}

func TestCleanupExpiredSinkService_WithValidJobValidBackup(t *testing.T) {
	httpMockHandler.Start()
	defer httpMockHandler.Stop()

	ctx := context.Background()

	configProvider := &MockImpersonatedTokenConfigProvider{
		TargetPrincipal: "backup-tooling@local-test-prod.iam.gserviceaccount.com",
		Error:           nil,
	}

	backupRepository, err := repository.NewBackupRepository(context.Background(), secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "backupRepository should be instantiate")

	service, err := newCleanupExpiredSinkService(context.Background(), configProvider, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "CleanupBackupService should be instantiate")

	deletedTargetBackupID := cleanupServiceBackupID + "-deleted"
	expiredTargetBackupID := cleanupServiceBackupID + "-expired"
	nonTargetBackupID := cleanupServiceBackupID + "-should-not-be-cleaned"
	_, err = backupRepository.AddBackup(ctx, cleanupBackupServiceBackup(deletedTargetBackupID, repository.BackupDeleted))
	require.NoError(t, err)

	backup := cleanupBackupServiceBackup(expiredTargetBackupID, repository.Finished)
	backup.SnapshotOptions = repository.SnapshotOptions{LifetimeInDays: 1}
	backup.EntityAudit = repository.EntityAudit{CreatedTimestamp: time.Now().Add(-48 * time.Hour)}
	_, err = backupRepository.AddBackup(ctx, backup)
	require.NoError(t, err)

	_, err = backupRepository.AddBackup(ctx, cleanupBackupServiceBackup(nonTargetBackupID, repository.Prepared))
	require.NoError(t, err)

	defer func() {
		deleteBackup(expiredTargetBackupID)
		deleteBackup(deletedTargetBackupID)
		deleteBackup(nonTargetBackupID)
	}()

	_, stdErr, err := captureStderr(func() {
		service.Run(ctx)
	})

	require.NoError(t, err)
	logMsg := "Cleaning up 1 sinks for type BigQuery"
	assert.Containsf(t, strings.TrimSpace(stdErr), logMsg, "Run should write log message %q but it logged\n\t%s", logMsg, stdErr)

	updatedDeletedBackup, err := backupRepository.GetBackup(ctx, deletedTargetBackupID)
	require.NoError(t, err)
	updatedExpiredBackup, err := backupRepository.GetBackup(ctx, expiredTargetBackupID)
	require.NoError(t, err)

	assert.Equalf(t, repository.BackupDeleted, updatedDeletedBackup.Status, "Backup with id %q should be be cleaned up but has status %s", deletedTargetBackupID, updatedDeletedBackup.Status)
	assert.Equalf(t, repository.BackupDeleted, updatedExpiredBackup.Status, "Backup with id %q should be be cleaned up but has status %s", expiredTargetBackupID, updatedExpiredBackup.Status)

	storageService, err := service2.NewStorageService(ctx, secret.NewEnvSecretProvider())

	require.NoError(t, err)

	count, err := storageService.DB().Model((*repository.SourceMetadata)(nil)).WhereIn("backup_id in (?)", []string{deletedTargetBackupID, expiredTargetBackupID}).Count()
	require.NoError(t, err)

	assert.Equal(t, 0, count)
}

func TestCleanupExpiredSinkService_WithExpiredBigQueryMirrorRevisions(t *testing.T) {
	httpMockHandler.Start()
	defer httpMockHandler.Stop()

	ctx := context.Background()

	configProvider := &MockImpersonatedTokenConfigProvider{
		TargetPrincipal: "backup-tooling@local-test-prod.iam.gserviceaccount.com",
		Error:           nil,
	}

	defer func() {
		dropSourceMetadata(cleanupServiceBackupID)
		dropJobs(cleanupServiceBackupID)
		deleteBackup(cleanupServiceBackupID)
	}()

	backupRepository, err := repository.NewBackupRepository(ctx, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "backupRepository should be instantiate")

	jobRepository, err := repository.NewJobRepository(ctx, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "jobRepository should be instantiate")

	sourceMetadataRepository, err := repository.NewSourceMetadataRepository(ctx, secret.NewEnvSecretProvider())
	require.NoError(t, err, "sourceMetadataRepository should be instantiate")

	sourceMetadataJobRepository, err := repository.NewSourceMetadataJobRepository(ctx, secret.NewEnvSecretProvider())
	require.NoError(t, err, "sourceMetadataRepository should be instantiate")

	service, err := newCleanupExpiredSinkService(ctx, configProvider, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "CleanupBackupService should be instantiate")

	backup := cleanupBackupServiceBackup(cleanupServiceBackupID, repository.Prepared)
	backup.Strategy = repository.Mirror
	_, err = backupRepository.AddBackup(ctx, backup)
	require.NoError(t, err, "should add new backup")

	nonTargetJob := cleanupServiceJobID + "-nonTarget"
	err = jobRepository.AddJob(ctx, cleanupServiceJob(nonTargetJob, repository.FinishedOk))
	require.NoError(t, err, "should add new job")

	targetJob := "daac5314-0472-4bf4-952c-7c418d4ef4f3"
	job := cleanupServiceJob(targetJob, repository.FinishedOk)
	job.Source = backup.Table[0]
	err = jobRepository.AddJob(ctx, job)
	require.NoError(t, err, "should add new job")

	sourceMetadataNonTarget, err := sourceMetadataRepository.Add(ctx, []*repository.SourceMetadata{
		{BackupID: cleanupServiceBackupID, Source: "amount_budget_plan_" + nonTargetJob, SourceChecksum: "checksum1", Operation: "Add"},
	})
	require.NoError(t, err, "sourceMetadata should be added")
	for _, m := range sourceMetadataNonTarget {
		err = sourceMetadataJobRepository.Add(ctx, m.ID, nonTargetJob)
		require.NoError(t, err)
	}

	sourceMetadatTarget, err := sourceMetadataRepository.Add(ctx, []*repository.SourceMetadata{
		{BackupID: cleanupServiceBackupID, Source: job.Source, SourceChecksum: "checksum2", Operation: "Add"},
		{BackupID: cleanupServiceBackupID, Source: job.Source, SourceChecksum: "checksum2", Operation: "Delete"},
	})
	require.NoError(t, err, "sourceMetadata should be added")
	for _, m := range sourceMetadatTarget {
		if m.Operation == "Add" {
			err = sourceMetadataJobRepository.Add(ctx, m.ID, targetJob)
			require.NoError(t, err)
		}
	}

	storageService, err := service2.NewStorageService(ctx, secret.NewEnvSecretProvider())

	require.NoError(t, err)

	_, err = storageService.DB().Model(&repository.Job{}).
		Set("audit_updated_timestamp=NOW()-interval '1 week'*?", maxMirrorRevisionLifetimeInWeeks+1).
		Where("id in (?)", pg.In([]interface{}{nonTargetJob, targetJob})).
		Update()
	require.NoError(t, err)

	_, err = storageService.DB().Model(&repository.SourceMetadata{}).
		Set("audit_created_timestamp=NOW()-interval '1 week'*?", maxMirrorRevisionLifetimeInWeeks+1).
		Where("backup_id = ? ", cleanupServiceBackupID).
		Update()
	require.NoError(t, err)

	_, stdErr, err := captureStderr(func() {
		service.Run(ctx)
	})
	require.NoError(t, err)

	logMsg := "No backups to clean up for type BigQuery"
	assert.Containsf(t, strings.TrimSpace(stdErr), logMsg, "Run should write log message '%s' but it logged\n\t%s", logMsg, stdErr)
	logMsg = "deleted 1 objects for backupID"
	assert.Containsf(t, strings.TrimSpace(stdErr), logMsg, "Run should write log message '%s' but it logged\n\t%s", logMsg, stdErr)

	count, err := storageService.DB().Model((*repository.SourceMetadata)(nil)).
		Where("id = ?", sourceMetadataNonTarget[0].ID).
		Where("audit_deleted_timestamp IS NULL").
		Count()
	require.NoError(t, err)

	assert.Equal(t, 1, count)

	count, err = storageService.DB().Model((*repository.SourceMetadata)(nil)).
		Where("id in (?)", pg.In([]interface{}{sourceMetadatTarget[0].ID, sourceMetadatTarget[1].ID})).
		Where("audit_deleted_timestamp IS NOT NULL").Count()
	require.NoError(t, err)
	assert.Equal(t, 2, count)

	updatedTargetJob, err := jobRepository.GetJob(ctx, targetJob)
	require.NoError(t, err)
	assert.Nil(t, updatedTargetJob)

	updatedNonTargetJob, err := jobRepository.GetJob(ctx, nonTargetJob)
	require.NoError(t, err)
	assert.NotNil(t, updatedNonTargetJob)
}

func TestCleanupExpiredSinkService_WithoutExpiredGcsMirrorRevisions(t *testing.T) {
	httpMockHandler.Start()
	defer httpMockHandler.Stop()

	ctx := context.Background()

	backupRepository, err := repository.NewBackupRepository(context.Background(), secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "backupRepository should be instantiate")

	service, err := newCleanupExpiredSinkService(context.Background(), nil, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "CleanupBackupService should be instantiate")

	backup := cleanupBackupServiceBackup(cleanupServiceBackupID, repository.Prepared)
	backup.Strategy = repository.Mirror
	_, err = backupRepository.AddBackup(ctx, backup)
	require.NoError(t, err, "should add new backup")
	defer func() { deleteBackup(cleanupServiceBackupID) }()

	_, stdErr, err := captureStderr(func() {
		service.Run(ctx)
	})

	require.NoError(t, err)
	logMsg := "[START] Deleting old CloudStorage revision"
	assert.NotContainsf(t, strings.TrimSpace(stdErr), logMsg, "Run should not write log message %q but it logged\n\t%s", logMsg, stdErr)
}

//this test is dependent on the grpc client of cloud storage which currently can't be mocked
func TestCleanupExpiredSinkService_WithExpiredGcsMirrorRevisions(t *testing.T) {
	t.Skip()

	httpMockHandler.Start()
	defer httpMockHandler.Stop()

	configProvider := &MockImpersonatedTokenConfigProvider{
		TargetPrincipal: "backup-tooling@local-test-prod.iam.gserviceaccount.com",
		Error:           nil,
	}

	httpMockHandler.Register(mock.ImpersonationHTTPMock)

	ctx := context.Background()
	backupRepository, err := repository.NewBackupRepository(ctx, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "backupRepository should be instantiate")

	sourceTrashcanRepository, err := repository.NewSourceTrashcanRepository(ctx, secret.NewEnvSecretProvider())
	require.NoError(t, err, "sourceMetadataRepository should be instantiate")

	service, err := newCleanupExpiredSinkService(ctx, configProvider, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "CleanupBackupService should be instantiate")

	backup := cleanupBackupServiceBackup(cleanupServiceBackupID, repository.Prepared)
	backup.Strategy = repository.Mirror
	backup.Type = repository.CloudStorage
	backup.CreatedTimestamp = time.Date(2019, 3, 14, 14, 57, 00, 0, time.Local)
	backup.CloudStorageOptions = repository.CloudStorageOptions{Bucket: "greg_test_bucket"}

	_, err = backupRepository.AddBackup(ctx, backup)
	require.NoError(t, err, "should add new backup")
	defer func() { deleteBackup(cleanupServiceBackupID) }()

	sourceTrashcanRepository.Add(ctx, cleanupServiceBackupID, "test_storage/2/10.txt", time.Now().AddDate(0, -1, 0))

	_, stdErr, err := captureStderr(func() {
		service.Run(ctx)
	})

	require.NoError(t, err)
	// FIXME currently the tests failed because we didn't find any way to mock GRPCConn, thus this tests checks for the failing error message
	// FIXME: any code which changes this behavior must be aware of this or rollback this bypass
	//logMsg := "[START] Deleting old CloudStorage revision"
	logMsg := "Expected OAuth 2 access token, login cookie or other valid authentication credential."
	assert.Containsf(t, strings.TrimSpace(stdErr), logMsg, "Run should not write log message %q but it logged\n\t%s", logMsg, stdErr)

	//var sources []repository.SourceTrashcan
	//exists, err := sourceTrashcanRepository.FilterExistingEntries(ctx, sources)
	//require.NoError(t, err)
	//assert.Len(t, exists, 0)
}

//this test is dependent on the grpc client of cloud storage which currently can't be mocked
func TestCleanupExpiredSinkService_WithTrashcanedGcsMirrorRevisions(t *testing.T) {
	t.Skip()

	httpMockHandler.Start()
	defer httpMockHandler.Stop()

	configProvider := &MockImpersonatedTokenConfigProvider{
		TargetPrincipal: "backup-tooling@local-test-prod.iam.gserviceaccount.com",
		Error:           nil,
	}

	httpMockHandler.Register(mock.ImpersonationHTTPMock)
	ctx := context.Background()

	backupRepository, err := repository.NewBackupRepository(ctx, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "backupRepository should be instantiate")

	service, err := newCleanupExpiredSinkService(ctx, configProvider, secret.NewEnvSecretProvider())
	require.NoErrorf(t, err, "CleanupBackupService should be instantiate")

	backup := cleanupBackupServiceBackup(cleanupServiceBackupID, repository.Prepared)
	backup.Strategy = repository.Mirror
	backup.Type = repository.CloudStorage
	backup.CreatedTimestamp = time.Date(2019, 3, 14, 14, 57, 00, 0, time.Local)
	backup.CloudStorageOptions = repository.CloudStorageOptions{Bucket: "greg_test_bucket"}

	_, err = backupRepository.AddBackup(ctx, backup)
	require.NoError(t, err, "should add new backup")
	defer func() { deleteBackup(cleanupServiceBackupID) }()

	_, stdErr, err := captureStderr(func() {
		service.Run(ctx)
	})

	require.NoError(t, err)
	// FIXME currently the tests failed because we didn't find any way to mock GRPCConn, thus this tests checks for the failing error message
	// FIXME: any code which changes this behavior must be aware of this or rollback this bypass
	//logMsg :="[START] Deleting old CloudStorage revision"
	logMsg := "Expected OAuth 2 access token, login cookie or other valid authentication credential."
	assert.Containsf(t, strings.TrimSpace(stdErr), logMsg, "Run should not write log message %q but it logged\n\t%s", logMsg, stdErr)

	//sourceTrashcanRepository, err := repository.NewSourceTrashcanRepository(ctx)
	//require.NoError(t, err, "sourceMetadataRepository should be instantiate")
	//var sources []repository.SourceTrashcan
	//sources = append(sources, repository.SourceTrashcan{BackupID: cleanupServiceBackupID, Source: "test_storage/2/10.txt"})
	//exists, err := sourceTrashcanRepository.FilterExistingEntries(ctx, sources)
	//require.NoError(t, err)
	//assert.Len(t, exists, 1)
}

func cleanupBackupServiceBackup(id string, status repository.BackupStatus) *repository.Backup {
	return &repository.Backup{
		ID:            id,
		Status:        status,
		SourceProject: "local-ability",
		Strategy:      repository.Snapshot,
		Type:          repository.BigQuery,
		SinkOptions: repository.SinkOptions{
			TargetProject: "local-ability-backup",
			Sink:          id + "-sink",
			Region:        "europe-west1",
			StorageClass:  repository.Nearline.String(),
		},
		BackupOptions: repository.BackupOptions{
			BigQueryOptions: repository.BigQueryOptions{"demo_delete_me_backup_target", []string{"gcp_billing_budget_amount_plan"}, []string{}},
		},
		EntityAudit: repository.EntityAudit{
			CreatedTimestamp: time.Now(),
		},
	}
}

func cleanupServiceJob(id string, status repository.JobStatus) *repository.Job {
	return &repository.Job{
		ID:       id,
		Source:   "amount_budget_plan_" + id,
		Status:   status,
		BackupID: cleanupServiceBackupID,
		Type:     repository.BigQuery,
	}
}
